Move address test helpers
