+++ pkg/kademlia/pslice/pslice_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/ethersphere/bee/pkg/kademlia/pslice"
 	"github.com/ethersphere/bee/pkg/swarm"
-	"github.com/ethersphere/bee/pkg/topology/test"
+	"github.com/ethersphere/bee/pkg/swarm/test"
 )
 
 // TestShallowestEmpty tests that ShallowestEmpty functionality works correctly.
